# perf: Use bit-reversed CRC32 computation with hash/crc32 package